Add mandatory legal footer for German signatures
Add an optional two line footer to the generated signature. Add a checkbox
to toggle adding a legal footer for employees in Germany on and off.

This legal footer is required to be part of all business communication in
Germany (see §35a GmbHG "Angaben auf Geschäftsbriefen"). It contains
information about the address of the company (first line of the footer)
and information about the incorporation of the company (second line).

A short info label informs the user, that this function is only relevant for
business communication in Germany.

This commit is missing the translations for Portuguese and Chinese.
These will be added soon. Please also note that the content of the
footer must not be translated as the German law requires it to be in German.

For other countries the footer functionality could be reused to add
non-legal information about the company, such a information about
company culture or the like.

modified:   public/javascript/properties.js
   - Add footer content in German to all languages
   - Add display strings to English, German and Spanish

new file:   public/jsx/checkbox.js
   - Add React component for checkboxes

modified:   public/jsx/email_signature_generator.js
   - Add check box for toggling German legal footer
   - Add info label
   - Add necessary checkbox change handling code

modified:   public/jsx/signature.js
   - Add option for adding two line footer to signature